11905: init commit, image difference model

+++ clearcut_detection_backend/sentinel_download.py
@@ -1,4 +1,6 @@
 import os
+import csv
+import datetime
 import subprocess
 
 from django.conf import settings
@@ -18,13 +20,13 @@ class SentinelDownload:
     def __init__(self):
         os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = './key.json'
         self.tile_dates_count = settings.MAXIMUM_DATES_REVIEWED_FOR_TILE
+        self.sequential_dates_count = settings.MAXIMUM_DATES_STORE_FOR_TILE
         self.storage_client = storage.Client()
         self.storage_bucket = self.storage_client.get_bucket('gcp-public-data-sentinel-2')
         self.config = ConfigParser(allow_no_value=True)
         self.config.read('gcp_config.ini')
         self.area_tile_set = self.config.get('config', 'AREA_TILE_SET').split()
         self.bands_to_download = self.config.get('config', 'BANDS_TO_DOWNLOAD').split()
-        self.resolution = self.config.get('config', 'RESOLUTION')
         self.executor = ThreadPoolExecutor(max_workers=10)
 
     def process_download(self):
@@ -63,13 +65,31 @@ class SentinelDownload:
         Iterates over folders to fetch tile images folder
         Downloads band specified in config
         :param tile_name:
-        :param tile_path:
+        :param tile_info:
         :return:
         """
-        tile_prefix = f'{tile_path}/IMG_DATA/R{self.resolution}/'
+        tile_prefix = f'{tile_path}/IMG_DATA/R20m/'
         blobs = self.storage_bucket.list_blobs(prefix=tile_prefix)
-        # check_blob = lambda value: any([value.endswith(f'_{band}_{self.resolution}.jp2')
-        #                                 for band in self.bands_to_download])
+
+        for blob in blobs:
+            band, download_needed = self.file_need_to_be_downloaded(blob.name)
+            if download_needed:
+                filename = os.path.join(DOWNLOADED_IMAGES_DIR, f'{tile_name}_{band}.jp2')
+                self.download_file_from_storage(blob, filename)
+                tile_info = TileInformation.objects.get(tile_name=tile_name)
+                if band == Bands.B11.value:
+                    tile_info.source_b11_location = filename
+                elif band == Bands.B12.value:
+                    tile_info.source_b12_location = filename
+                elif band == Bands.B8A.value:
+                    tile_info.source_b8a_location = filename
+                else:
+                    continue
+                tile_info.save()
+        
+        tile_prefix = f'{tile_path}/IMG_DATA/R10m/'
+        blobs = self.storage_bucket.list_blobs(prefix=tile_prefix)
+
         for blob in blobs:
             band, download_needed = self.file_need_to_be_downloaded(blob.name)
             if download_needed:
@@ -83,8 +103,20 @@ class SentinelDownload:
                 elif band == Bands.TCI.value:
                     tile_info.source_tci_location = filename
                 else:
-                    print('Tile with unknown band has been downloaded')
+                    continue
                 tile_info.save()
+
+        tile_prefix = f'{tile_path}/QI_DATA/' #MSK_CLDPRB_20m.jp2
+        blobs = self.storage_bucket.list_blobs(prefix=tile_prefix)
+        for blob in blobs:
+            if blob.name.endswith('MSK_CLDPRB_20m.jp2'):
+                tile_info = TileInformation.objects.get(tile_name=tile_name)
+                filename = os.path.join(DOWNLOADED_IMAGES_DIR, f"{tile_name}_{blob.name.split('/')[-1]}")
+                self.download_file_from_storage(blob, filename)
+                tile_info.source_clouds_location = filename
+                tile_info.save()
+        #print(TileInformation.objects.values())
+        #print(list(TileInformation.objects.values('tile_index').distinct()))
 
     def file_need_to_be_downloaded(self, name):
         """
@@ -93,7 +125,7 @@ class SentinelDownload:
         :return:
         """
         for band in self.bands_to_download:
-            if name.endswith(f'_{band}_{self.resolution}.jp2'):
+            if name.endswith(f'_{band}_10m.jp2') or name.endswith(f'_{band}_20m.jp2'):
                 return band, True
         return None, False
 
@@ -114,26 +146,31 @@ class SentinelDownload:
             command = f'gsutil ls -l {base_uri}/{tile_uri}/ | sort -k2n | tail -n{self.tile_dates_count}'
             granule_id_list = self.find_granule_id(command)
             granule_id_list.reverse()
+            granule_num = 0
             for granule_id in granule_id_list:
-                print('====== GRANULE ID ======')
-                print(granule_id)
-                nested_command = f'gsutil ls -l {base_uri}/{tile_uri}/{granule_id}/GRANULE/ | sort -k2n | tail -n1'
-                nested_granule_id_list = self.find_granule_id(nested_command)
-                nested_granule_id = nested_granule_id_list[0]
-                updated_tile_uri = f'{tile_uri}/{granule_id}/GRANULE/{nested_granule_id}'
-                filename = os.path.join(DOWNLOADED_IMAGES_DIR, f'{tile_name}_{metadata_file}')
-                try:
-                    blob = self.storage_bucket.get_blob(f'{updated_tile_uri}/{metadata_file}')
-                    update_needed = self.define_if_tile_update_needed(blob, tile_name, filename)
-                    print('====== IS UPDATE NEEDED ======')
-                    print(update_needed)
-                    if update_needed:
-                        tiles_to_be_downloaded[f'{tile_name}'] = f'{updated_tile_uri}'
-                        os.remove(filename)
-                except Exception as e:
-                    print(e)
-                    print(dir(e))
-
+                if granule_num < self.sequential_dates_count:
+                    print('====== GRANULE ID ======')
+                    print(granule_id)
+                    nested_command = f'gsutil ls -l {base_uri}/{tile_uri}/{granule_id}/GRANULE/ | sort -k2n | tail -n1'
+                    nested_granule_id_list = self.find_granule_id(nested_command)
+                    nested_granule_id = nested_granule_id_list[0]
+                    updated_tile_uri = f'{tile_uri}/{granule_id}/GRANULE/{nested_granule_id}'
+                    filename = os.path.join(DOWNLOADED_IMAGES_DIR, f'{tile_name}_{metadata_file}')
+                    try:
+                        blob = self.storage_bucket.get_blob(f'{updated_tile_uri}/{metadata_file}')
+                        update_needed = self.define_if_tile_update_needed(blob, f'{tile_name}_{granule_num}', filename)
+                        print('====== IS UPDATE NEEDED ======')
+                        print(update_needed)
+                        if update_needed:
+                            tiles_to_be_downloaded[f'{tile_name}_{granule_num}'] = f'{updated_tile_uri}'
+                            os.remove(filename)
+                            granule_num+=1
+                    except Exception as e:
+                        print(e)
+                        print(dir(e))
+                else:
+                    break
+        print(tiles_to_be_downloaded)
         return tiles_to_be_downloaded
 
     def define_if_tile_update_needed(self, blob, tile_name, filename) -> bool:
@@ -146,7 +183,9 @@ class SentinelDownload:
         :param filename:
         :return:
         """
+        
         tile_info, created = TileInformation.objects.get_or_create(tile_name=tile_name)
+
         if not created:
             update_needed = blob.md5_hash != tile_info.tile_metadata_hash
             if not update_needed:
@@ -164,6 +203,8 @@ class SentinelDownload:
         if cloud_coverage_value <= settings.MAXIMUM_CLOUD_PERCENTAGE_ALLOWED:
             tile_info.cloud_coverage = cloud_coverage_value
             tile_info.tile_metadata_hash = blob.md5_hash
+            tile_info.tile_index = tile_name.split('_')[0]
+            tile_info.tile_date = datetime.datetime.strptime(blob.name.split('_')[-2][:8], '%Y%m%d')
             tile_info.save()
             return True
         else:
